# backend/api.yaml
openapi: "3.1.0"
info:
  title: "backend API"
  description: "backend API"
  version: "1.0.0"
servers:
  - url: "https://backend"
paths:
...
  /api/appointment/{id}:
    put:
      summary: "PUT api/appointment/{id}"
      operationId: "updateAppointment"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Appointment"
    delete:
      summary: "DELETE api/appointment/{id}"
      operationId: "deleteAppointment"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Void"
    get:
      summary: "GET api/appointment/{id}"
      operationId: "getAppointmentById"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Appointment"
  /api/appointment:
    post:
      summary: "POST api/appointment"
      operationId: "createAppointment"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Appointment"
    get:
      summary: "GET api/appointment"
      operationId: "getAllAppointments"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Appointment"
  /api/appointment/patient:
    post:
      summary: "POST api/appointment/patient"
      operationId: "createAppointmentByPatient"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentCreation"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Appointment"
  /api/appointment/patient/{patientId}:
    get:
      summary: "GET api/appointment/patient/{patientId}"
      operationId: "findAppointmentsByPatient"
      parameters:
        - name: "patientId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Appointment"
  /api/appointment/schedule/{scheduleId}:
    get:
      summary: "GET api/appointment/schedule/{scheduleId}"
      operationId: "findAppointmentsBySchedule"
      parameters:
        - name: "scheduleId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Appointment"
  /api/appointment/doctor/{doctorId}:
    get:
      summary: "GET api/appointment/doctor/{doctorId}"
      operationId: "findAppointsByDoctor"
      parameters:
        - name: "doctorId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Appointment"
  /api/doctor/{id}:
    put:
      summary: "PUT api/doctor/{id}"
      operationId: "updateDoctor"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DoctorRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Doctor"
    delete:
      summary: "DELETE api/doctor/{id}"
      operationId: "deleteDoctor"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Void"
    get:
      summary: "GET api/doctor/{id}"
      operationId: "getDoctorById"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Doctor"
...
  /api/patient/{id}:
    put:
      summary: "PUT api/patient/{id}"
      operationId: "updatePatient"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Patient"
    delete:
      summary: "DELETE api/patient/{id}"
      operationId: "deletePatient"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Void"
    get:
      summary: "GET api/patient/{id}"
      operationId: "getPatientById"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Patient"
  /api/patient:
    post:
      summary: "POST api/patient"
      operationId: "createPatient"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatientRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Patient"
    get:
      summary: "GET api/patient"
      operationId: "getAllPatients"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Patient"
...
  /api/diagnosis/{id}:
    put:
      summary: "PUT api/diagnosis/{id}"
      operationId: "updateDiagnosis"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagnosisRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Diagnosis"
    delete:
      summary: "DELETE api/diagnosis/{id}"
      operationId: "deleteDiagnosis"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Void"
    get:
      summary: "GET api/diagnosis/{id}"
      operationId: "getDiagnosisById"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Diagnosis"
  /api/diagnosis:
    post:
      summary: "POST api/diagnosis"
      operationId: "createDiagnosis"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiagnosisRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Diagnosis"
    get:
      summary: "GET api/diagnosis"
      operationId: "getAllDiagnosis"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/Diagnosis"
...
  /api/admission/appointment/{appointmentId}:
    get:
      summary: "GET api/admission/appointment/{appointmentId}"
      operationId: "findAdmissionByAppointment"
      parameters:
        - name: "appointmentId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Admission"
...
  /api/schedule/{id}:
    put:
      summary: "PUT api/schedule/{id}"
      operationId: "updateSchedule"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleRequest"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Schedule"
    delete:
      summary: "DELETE api/schedule/{id}"
      operationId: "deleteSchedule"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Void"
    get:
      summary: "GET api/schedule/{id}"
      operationId: "getScheduleById"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Schedule"
...
components:
  schemas:
    AppointmentRequest:
      type: "object"
      properties:
        date:
          type: "string"
          format: "date"
        time:
          type: "string"
          format: "partial-time"
        queueNo:
          type: "integer"
          format: "int32"
        status:
          type: "string"
          enum:
            - "PENDING"
        patientId:
          type: "integer"
          format: "int64"
        scheduleId:
          type: "integer"
          format: "int64"
    Patient:
      type: "object"
      properties:
        emergencyContactNo:
          type: "string"
        personId:
          type: "integer"
          format: "int64"
        dob:
          type: "string"
          format: "date"
        name:
          type: "string"
        address:
          type: "string"
        gender:
          type: "string"
        email:
          type: "string"
        phoneNo:
          type: "string"
        role:
          type: "string"
          enum:
            - "USER"
            - "STAFF"
            - "ADMIN"
    Department:
      type: "object"
      properties:
        departmentId:
          type: "integer"
          format: "int64"
        name:
          type: "string"
    Ward:
      type: "object"
      properties:
        wardId:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        capacity:
          type: "integer"
          format: "int32"
        location:
          type: "string"
    Doctor:
      type: "object"
      properties:
        speciality:
          type: "string"
        regNo:
          type: "string"
        department:
          $ref: "#/components/schemas/Department"
        ward:
          $ref: "#/components/schemas/Ward"
        personId:
          type: "integer"
          format: "int64"
        dob:
          type: "string"
          format: "date"
        name:
          type: "string"
        address:
          type: "string"
        gender:
          type: "string"
        email:
          type: "string"
        phoneNo:
          type: "string"
        role:
          type: "string"
          enum:
            - "USER"
            - "STAFF"
            - "ADMIN"
    Schedule:
      type: "object"
      properties:
        scheduleId:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        description:
          type: "string"
        startTime:
          type: "string"
          format: "partial-time"
        endTime:
          type: "string"
          format: "partial-time"
        day:
          type: "string"
          enum:
            - "SUNDAY"
            - "MONDAY"
            - "TUESDAY"
            - "WEDNESDAY"
            - "THURSDAY"
            - "FRIDAY"
            - "SATURDAY"
        doctor:
          $ref: "#/components/schemas/Doctor"
    Appointment:
      type: "object"
      properties:
        appointmentId:
          type: "integer"
          format: "int64"
        date:
          type: "string"
          format: "date"
        time:
          type: "string"
          format: "partial-time"
        queueNo:
          type: "integer"
          format: "int32"
        status:
          type: "string"
          enum:
            - "PENDING"
        patient:
          $ref: "#/components/schemas/Patient"
        schedule:
          $ref: "#/components/schemas/Schedule"
    AppointmentCreation:
      type: "object"
      properties:
        date:
          type: "string"
          format: "date"
        scheduleId:
          type: "integer"
          format: "int64"
    Void:
      type: "object"
      properties: { }
    DoctorRequest:
      type: "object"
      properties:
        speciality:
          type: "string"
        regNo:
          type: "string"
        departmentId:
          type: "integer"
          format: "int64"
        wardId:
          type: "integer"
          format: "int64"
        dob:
          type: "string"
          format: "date"
        name:
          type: "string"
        address:
          type: "string"
        email:
          type: "string"
        gender:
          type: "string"
        phoneNo:
          type: "string"
        password:
          type: "string"
...
    PatientRequest:
      type: "object"
      properties:
        emergencyContactNo:
          type: "string"
        dob:
          type: "string"
          format: "date"
        name:
          type: "string"
        address:
          type: "string"
        email:
          type: "string"
        gender:
          type: "string"
        phoneNo:
          type: "string"
        password:
          type: "string"
...
    MedicalService:
      type: "object"
      properties:
        serviceId:
          type: "integer"
          format: "int64"
        name:
          type: "string"
        description:
          type: "string"
        category:
          type: "string"
        price:
          type: "number"
          format: "float"
...
    DiagnosisRequest:
      type: "object"
      properties:
        admissionId:
          type: "integer"
          format: "int64"
        diagnosis:
          type: "string"
        prescription:
          type: "string"
    Report:
      type: "object"
      properties:
        reportId:
          type: "integer"
          format: "int64"
        result:
          type: "string"
        status:
          type: "string"
          enum:
            - "PENDING"
            - "COMPLETED"
        medicalService:
          $ref: "#/components/schemas/MedicalService"
        admission:
          $ref: "#/components/schemas/Admission"
    Admission:
      type: "object"
      properties:
        admissionId:
          type: "integer"
          format: "int64"
        admissionDate:
          type: "string"
          format: "date"
        dischargeDate:
          type: "string"
          format: "date"
        description:
          type: "string"
        appointment:
          $ref: "#/components/schemas/Appointment"
        diagnosis:
          $ref: "#/components/schemas/Diagnosis"
        reports:
          type: "array"
          items:
            $ref: "#/components/schemas/Report"
    Diagnosis:
      type: "object"
      properties:
        diagnosisId:
          type: "integer"
          format: "int64"
        admission:
          $ref: "#/components/schemas/Admission"
        diagnosis:
          type: "string"
        prescription:
          type: "string"
...
    ScheduleRequest:
      type: "object"
      properties:
        name:
          type: "string"
        description:
          type: "string"
        startTime:
          type: "string"
          format: "partial-time"
        endTime:
          type: "string"
          format: "partial-time"
        day:
          type: "string"
          enum:
            - "SUNDAY"
            - "MONDAY"
            - "TUESDAY"
            - "WEDNESDAY"
            - "THURSDAY"
            - "FRIDAY"
            - "SATURDAY"
        doctorId:
          type: "integer"
          format: "int64"
...